requirements.txt
openai
colorama
json
tiktoken
python-dotenv